LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/webmessaging/broadcastchannel/opaque-origin-expected.txt

CONSOLE MESSAGE: NotSupportedError: Shared workers are not supported

Harness Error (FAIL), message = NotSupportedError: Shared workers are not supported

PASS Opaque origin should be serialized to "null"
PASS BroadcastChannel messages from opaque origins should be self-contained
PASS BroadcastChannel messages from data URL dedicated workers should be self-contained
TIMEOUT BroadcastChannel messages from data URL shared workers should be self-contained Test timed out

Harness Error (FAIL), message = NotSupportedError: Shared workers are not supported

PASS Opaque origin should be serialized to "null"
PASS BroadcastChannel messages from opaque origins should be self-contained
PASS BroadcastChannel messages from data URL dedicated workers should be self-contained
TIMEOUT BroadcastChannel messages from data URL shared workers should be self-contained Test timed out
